Fix rev_1_2_3 test
The test was failing because we declared the wrong expected value.
The expected value we declared made use of
`nil = (vcon (PolymorphicList Nat) 0)` and
`cons = (vcon (PolymorphicList Nat) 1)`, which are not normal forms.
Since we do _not_ evaluate the expected value, the expected value
must already be in normal form.
Since this was not the case, the expected value was not identical to
the actual value (which _was_ in normal form).