Use "ok" and "ng" instead of "legal" and "illegal"
The word "legal" and "illegal" look too similar.
When I'm skimming the (usually long) test names, I want to
be able to distinguish legal tests
from illegal tests in a split-second.
The similarity between the two words makes this hard.

So, I use "ok" instead of "legal",
and "ng" (short for "no good") instead of "illegal".