Return toprightmost substitutable defs in `may_to_zo`
Explanation of "toprightmost defs":

Suppose we have the following code:

```may
ind Bool
    case false
    case true
return Set0

// Overly complicated definition of `and` that uses DeMorgan's law.
fun and(x: Bool, y: Bool): Bool
    fun not(a: Bool): Bool
        match a
        case false:
            true
        case true:
            false
        return1 Bool

    fun or(a: Bool, b: Bool): Bool
        match a
        case false:
            b
        case true:
            true
        return1 Bool

    // By DeMorgan's law:
    not(or(not(x), not(y)))

and(true, true)
```

In the above example, we can draw an AST like this
(details are omitted for brevity):

```text
                   [ind Bool]
                      \
                       \
                      [fun and]
                      /      \
                     /        \
                  [fun not]    \
                  /  \          \
                 /    \        {and(true, true)}
         {match...}   [fun or]
                      /      \
               {match...}     {not(or(not(x), not(y)))}
```

In the above AST, observe that:

Each definition
(i.e., `ind Bool`, `fun and`, `fun not`, and `fun or`) is a node.
The definition nodes are enclosed in square brackets ("[]").

Each definition's value (traditionally called the "right hand side")
is said definition's left child.

Each definition's dependency is said definition's right child.

Leaf values are enclosed in curly braces ("{}").

The "toprightmost defs" are `ind Bool` and `fun and`.

The definitions `fun not` and `fun or` are **not** toprightmost.